backend/requirements.txt
fastapi
uvicorn
httpx
asyncio
googlesearch-python